renamed main file and visualization and hyperparameter tuning and lemmatization
